+++ h5rdmtoolbox/ld/hdf/groups.py
@@ -6,14 +6,14 @@ from ontolutils.namespacelib.hdf5 import HDF5
 from rdflib import Namespace
 from rdflib import RDF
 
-from h5rdmtoolbox.wrapper.ld.hdf.attributes import process_attribute
-from h5rdmtoolbox.wrapper.ld.hdf.datasets import process_dataset
-from h5rdmtoolbox.wrapper.ld.utils import get_obj_bnode
+from h5rdmtoolbox.ld.hdf.attributes import process_attribute
+from h5rdmtoolbox.ld.hdf.datasets import process_dataset
+from h5rdmtoolbox.ld.utils import get_obj_bnode
 
 HDF = Namespace(str(HDF5))
 
 
-def process_group(group, graph, parent_uri, blank_node_iri_base: Optional[str] = None):
+def process_group(group, graph, parent_uri, blank_node_iri_base: Optional[str] = None, skipND: int = 1):
     """Recursively process HDF5 groups and datasets, adding them to the RDF graph."""
     group_uri = get_obj_bnode(group, blank_node_iri_base=blank_node_iri_base)
     graph.add((group_uri, RDF.type, HDF.Group))
@@ -27,7 +27,7 @@ def process_group(group, graph, parent_uri, blank_node_iri_base: Optional[str] =
         if isinstance(item, h5py.Group):
             process_group(item, graph, group_uri, blank_node_iri_base=blank_node_iri_base)
         elif isinstance(item, h5py.Dataset):
-            process_dataset(item, graph, group_uri, item_uri, blank_node_iri_base=blank_node_iri_base)
+            process_dataset(item, graph, group_uri, item_uri, blank_node_iri_base=blank_node_iri_base, skipND=skipND)
 
     # Process attributes of the group
     for attr, value in group.attrs.items():