+++ h5rdmtoolbox/ld/user/datasets.py
@@ -1,9 +1,9 @@
 import rdflib
 
-from h5rdmtoolbox.wrapper.ld.user.attributes import process_attribute
-from h5rdmtoolbox.wrapper.ld.utils import get_obj_bnode
-from h5rdmtoolbox.wrapper.rdf import PROTECTED_ATTRIBUTE_NAMES
-
+from h5rdmtoolbox.ld.user.attributes import process_attribute
+from h5rdmtoolbox.ld.utils import get_obj_bnode
+from h5rdmtoolbox.ld.rdf import PROTECTED_ATTRIBUTE_NAMES
+from ..rdf import RDFManager
 
 def process_dataset(dataset, graph, blank_node_iri_base=None):
     dataset_uri = get_obj_bnode(dataset, blank_node_iri_base=blank_node_iri_base)
@@ -11,6 +11,6 @@ def process_dataset(dataset, graph, blank_node_iri_base=None):
         if ak not in PROTECTED_ATTRIBUTE_NAMES:
             process_attribute(dataset, ak, av, graph, blank_node_iri_base=blank_node_iri_base)
 
-    rdf_type = dataset.rdf.type
+    rdf_type = RDFManager(dataset.attrs).type
     if rdf_type:
         graph.add((dataset_uri, rdflib.RDF.type, rdflib.URIRef(rdf_type)))