+++ h5rdmtoolbox/wrapper/__init__.py
@@ -4,7 +4,5 @@ Contains wrapper classes
 """
 
 from . import core
-from . import lazy
-from .jsonld import hdf2jsonld
 
-__all__ = ['core', 'lazy', 'hdf2jsonld']
+__all__ = ['core']