+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 Log of changes in the versions
 
+## v1.7.2
+
+- remove `skipND` from being deprecated
+- separation of "linked data" code into a separate repo
+- update documentation
+
 ## v1.7.1
 
 - fix documentation at readthedocs